Add True Color support.
Konsole, st, iTerm2 support True Color in terminals.  This patch adds
support for it.  The old `render_rgb` is renamed to `render_256` because
it is limited to 256 colors.  The new default `render_rgb` uses the True
Color output.

There probably should be some terminfo support to determine what the
terminal actually supports.

Signed-off-by: Rüdiger Sonderfeld <ruediger@c-plusplus.de>